// mc_frontend/openapi/model/nodeMC.ts
/**
 * Machina Callida Backend REST API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NodeMC { 
    /**
     * Node name as given by ANNIS.
     */
    annis_node_name?: string;
    /**
     * Node type as given by ANNIS.
     */
    annis_node_type?: string;
    /**
     * Raw word form as given by ANNIS.
     */
    annis_tok?: string;
    /**
     * Node type as given by ANNIS (?).
     */
    annis_type?: string;
    /**
     * Unique identifier for the node in the SALT model.
     */
    id?: string;
    /**
     * Whether the raw word form is missing in a given vocabulary.
     */
    is_oov?: boolean;
    /**
     * Lemmatized word form.
     */
    udep_lemma?: string;
    /**
     * Universal part of speech tag for the word form.
     */
    udep_upostag?: string;
    /**
     * Language-specific part of speech tag for the word form.
     */
    udep_xpostag?: string;
    /**
     * Additional morphological information.
     */
    udep_feats?: string;
    /**
     * Solution value for this node in an exercise.
     */
    solution?: string;
}